Restaurant booking API
Through this api, restaurants can build their own dishes and manage the price, name and other details of each dish. Users can order meals through this API, and the model of the meal and the information of the dish created by the restaurant will be associated in the form of an external key in the database.